Service task: Minor Test stuff refactoring:  - The module `utils_test` has been introduced.  - The `AbstractTest` class used for many tests of the `hackerrank` module has been moved to `utils_test` in order to use it also in other modules.  - However wie still keep an `AbstractTest` class at the `hackerrank` module in order to avoid the necessity to update many test classes/files at `hackerrank`. It inherits the class moved to `utils_test` without introducing anything new.
